finish single queue refactor
