+++ pomf-pusher/app/GlobalPusher.scala
@@ -1,13 +1,14 @@
-package settings
-
 import play.api._
+import akka.actor._
 import akka.actor.ActorSystem
-import service.PomfNotificationService
+import service.PusherNotificationService
 
 object GlobalPusher extends GlobalSettings {
 
   implicit val system = ActorSystem("PusherSystem")
     
+  val pusherNotificationService = system.actorOf(Props[PusherNotificationService], "pusher-listenner-service")
+
   override def onStart(app: Application) {
     Logger.info("Pomf-pusher has started")
   }
